debian: Run as ndn user with its home directory
Create and run as an `ndn-user` user with HOME=/var/lib/ndn, the same as
in the ndn-cxx package. Keys will be installed to /var/lib/ndn/.ndn by
programs from both packages, and the key store will be shared.

As part of this, the nfd binary needs to have some additional
capabilities set, which adds a build-dep on libcap2-bin for setcap.

We can’t run the nfd daemon as ndn-user from the start (using
systemd’s User key) as it does its own socket setup, which requires root
privileges. The nfd configuration has been changed to drop privileges
and run as ndn-user after startup. nfd does not support systemd (or inetd)
socket activation.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T13932